add some comments + rewrite view script

+++ include/polygon2d.h
@@ -16,6 +16,12 @@ class Polygon2D
 
         size_t layer() const;
         void set_layer(const size_t layer);
+
+        /** layer value calculation
+         * calculation of the polygon layer based on the analysis 
+         * of intersections of rays emitted from each polygon point with other polygons
+         * @param `polygons` vector of polygon2D
+         */
         void calc_layer(const std::vector<Polygon2D>& polygons);
         
         std::vector<Segment2D>::const_iterator border_begin() const;
@@ -30,7 +36,16 @@ class Polygon2D
         bool operator!=(const Polygon2D& polygon);
     
     protected:
+        /** ray entry calculation
+         * calculation of ray entry into the polygon based 
+         * on its intersections with segments
+         * @param `ray` segment based on part of polygon and fixed remote point
+         * @param `polygon` polygon for calculating ray occurrences
+         */
         size_t calc_intersec(const Segment2D& ray, const Polygon2D& polygon);
-        float scalar_product(const Point2D& first, const Point2D& second);
+        /** scalar product for 2D points
+         * @param `first`, `second` two 2D points
+         */
+        double scalar_product(const Point2D& first, const Point2D& second);
 
 };
